+++ cloud/datasync/src/SyncServer.ts
@@ -4,16 +4,6 @@ import {setupGlobalHandlers, setupHandlers} from './crud-handlers/HandlerMapper'
 import SyncDataSetOptions from './options/SyncDatasetOptions';
 import SyncOptions from './options/SyncGlobalOptions';
 import SyncApi from './SyncApi';
-import {BunyanLogger,LOG_LEVEL} from '../../logger-cloud/src/index';
-//import {BunyanLogger,LOG_LEVEL} from '@raincatcher/logger-cloud'; this don't work
-
-// constructor for loggers
-const logD = new BunyanLogger(LOG_LEVEL.DEBUG);
-const logE = new BunyanLogger(LOG_LEVEL.ERROR);
-const logI = new BunyanLogger(LOG_LEVEL.INFO);
-const logL = new BunyanLogger(LOG_LEVEL.LOG);
-const logW = new BunyanLogger(LOG_LEVEL.WARN);
-
 
 /**
  * Implementation for sync server side api
@@ -27,12 +17,10 @@ export const SyncServer: SyncApi = {
    */
     connect(options: SyncOptions, callback: (err: any) => void) {
     if (options.globalSyncOptions) {
-      logI.info('SyncOptions set :', options.globalSyncOptions);
       sync.setConfig(options.globalSyncOptions);
     }
     const sdo = options.datasetConfiguration;
     sync.connect(sdo.mongoDbConnectionUrl, sdo.mongoDbOptions, sdo.redisConnectionUrl, function(err: any) {
-      logE.error("SyncServer sync.connect Error", err);
       callback(err);
     });
   },
@@ -52,18 +40,15 @@ export const SyncServer: SyncApi = {
     registerDatasetDataHandler(datasetId: string, options: SyncDataSetOptions, dataHandler?: DataSetHandler) {
     sync.init(datasetId, options, function(err: any) {
       if (err) {
-        logE.error("SyncServer sync.init Error", err);
         throw new Error(err);
       } else {
         // set optional custom collision handler if its a function
         if (options && options.collisionHandler) {
-          logI.info('Custom collision handler set', options.collisionHandler);
           sync.handleCollision(datasetId, options.collisionHandler);
         }
 
         // Set optional custom hash function to deal with detecting model changes.
         if (options && options.hashFunction) {
-          logI.info('Set hash function to deal with detecting model changes : ',options.hashFunction);
           sync.setRecordHashFn(datasetId, options.hashFunction);
         }
         if (dataHandler) {
