+++ cloud/logger-cloud/integration/index.ts
@@ -1,0 +1,4 @@
+
+// basic usage
+/*let logger = new BunyanLogger(LOG_LEVEL.DEBUG);
+ logger.debug("debug message ", {test:"test"});*/
