+++ cloud/logger-cloud/src/index.ts
@@ -1,9 +1,5 @@
-/**
- * Created by acunningham on 15/06/17.
- */
-
 import {BunyanLogger} from './BunyanLogger'
 
-export {Logger, LOG_LEVEL} from  './Logger';
+export {Logger} from  './Logger';
 export {BunyanLogger} from  './BunyanLogger';
